requirements.txt
flask
Flask-SQLAlchemy
flask-login
bs4
requests
gunicorn